requirements.txt
python-dotenv==0.21.0
requests==2.27.1
webencodings==0.5.1
xmltodict==0.13.0

